fix boat positions
