+++ index.js
@@ -1,6 +1,7 @@
 import { httpServer } from "./src/http_server/index.js";
 import { WebSocketServer } from "ws";
 import { parseMaybeJson, sendToAll, makeJsonMsg } from "./util.js";
+import { handleReg, handleCreateRoom, handleAddUserToRoom, handleAddShips, handleAttack } from "./wsHandlers.js";
 
 const HTTP_PORT = 8181;
 
@@ -15,6 +16,7 @@ let winners = [];
 let nextRoomId = 1;
 let nextGameId = 1;
 let nextPlayerIndex = 1;
+const nextPlayerIndexRef = { value: nextPlayerIndex };
 
 const WS_PORT = 3000;
 const wss = new WebSocketServer({ port: WS_PORT });
@@ -38,206 +40,23 @@ wss.on("connection", (ws) => {
 		}
 		console.log("Received:", cmd);
 		if (cmd.type === "reg") {
-			const data = parseMaybeJson(cmd.data);
-			const { name, password } = data;
-			let error = false;
-			let errorText = "";
-			let index;
-			if (!name || !password) {
-				error = true;
-				errorText = "Name and password required";
-			} else if (players.has(name)) {
-				if (players.get(name).password !== password) {
-					error = true;
-					errorText = "Incorrect password";
-				} else {
-					index = players.get(name).index;
-					players.get(name).ws = ws;
-				}
-			} else {
-				index = nextPlayerIndex++;
-				players.set(name, { password, wins: 0, index, ws });
-			}
-			const response = makeJsonMsg("reg", { name, index, error, errorText });
-			ws.send(response);
-			console.log("Sent:", JSON.parse(response));
-			sendUpdateRoom();
-			sendUpdateWinners();
+			handleReg(cmd, ws, players, sendUpdateRoom, sendUpdateWinners, nextPlayerIndexRef);
 		} else if (cmd.type === "create_room") {
-			let playerName, playerIndex;
-			for (const [name, info] of players.entries()) {
-				if (info.ws === ws) {
-					playerName = name;
-					playerIndex = info.index;
-					break;
-				}
-			}
-
-			if (!playerName) {
-				playerName = Array.from(players.keys()).pop();
-				playerIndex = players.get(playerName).index;
-				players.get(playerName).ws = ws;
-			}
-
-			const roomId = nextRoomId++;
-			rooms.set(roomId, {
-				roomId,
-				roomUsers: [{ name: playerName, index: playerIndex }],
-			});
-
-			sendUpdateRoom();
-
-			console.log("Room created:", { roomId, playerName, playerIndex });
-			console.log("All rooms:", Array.from(rooms.entries()));
+			handleCreateRoom(ws, players, rooms, () => nextRoomId++, sendUpdateRoom);
 		} else if (cmd.type === "add_user_to_room") {
-			const data = parseMaybeJson(cmd.data);
-			const { indexRoom } = data || {};
-			console.log("Trying to join room:", indexRoom);
-
-			let playerName, playerIndex;
-			for (const [name, info] of players.entries()) {
-				if (info.ws === ws) {
-					playerName = name;
-					playerIndex = info.index;
-					break;
-				}
-			}
-			if (!playerName) {
-				playerName = Array.from(players.keys()).pop();
-				playerIndex = players.get(playerName).index;
-				players.get(playerName).ws = ws;
-			}
-			console.log("Player joining room:", { playerName, playerIndex });
-
-			const roomId = Number(indexRoom);
-			const room = rooms.get(roomId);
-
-			console.log("Found room:", room);
-
-			if (room && room.roomUsers.length === 1) {
-				room.roomUsers.push({ name: playerName, index: playerIndex });
-
-				sendUpdateRoom();
-
-				const idGame = nextGameId++;
-				const [p1, p2] = room.roomUsers;
-				games.set(idGame, {
-					idGame,
-					players: [p1, p2],
-					ships: {},
-					turn: null,
-					finished: false,
-				});
-
-				room.roomUsers.forEach((u) => {
-					const wsTarget = players.get(u.name).ws;
-					if (wsTarget && wsTarget.readyState === 1) {
-						const resp = makeJsonMsg("create_game", { idGame, idPlayer: u.index });
-						wsTarget.send(resp);
-						console.log("Sent create_game to player:", u.name);
-					} else {
-						console.log("Warning: Could not send create_game to player:", u.name);
-					}
-				});
-			} else {
-				console.log("Failed to join room - Room not found or already full");
-			}
+			handleAddUserToRoom(cmd, ws, players, rooms, games, () => nextGameId++, sendUpdateRoom, makeJsonMsg);
 		} else if (cmd.type === "add_ships") {
-			const data = parseMaybeJson(cmd.data);
-			const { gameId, ships, indexPlayer } = data || {};
-			const game = games.get(Number(gameId));
-			if (!game) return;
-			game.ships[indexPlayer] = ships;
-			if (Object.keys(game.ships).length === 2) {
-				const [p1, p2] = game.players;
-				const currentPlayerIndex = Math.random() < 0.5 ? p1.index : p2.index;
-				game.turn = currentPlayerIndex;
-				game.players.forEach((u) => {
-					const wsTarget = players.get(u.name).ws;
-					if (wsTarget && wsTarget.readyState === 1) {
-						const resp = makeJsonMsg("start_game", {
-							ships: game.ships[u.index],
-							currentPlayerIndex,
-						});
-						wsTarget.send(resp);
-						console.log("Sent:", resp);
-					}
-				});
-				sendTurn(game);
-			}
+			handleAddShips(cmd, players, games, sendTurn, makeJsonMsg);
 		} else if (cmd.type === "attack" || cmd.type === "randomAttack") {
-			const data = parseMaybeJson(cmd.data);
-			const { gameId, x, y, indexPlayer } = data || {};
-			const game = games.get(Number(gameId));
-			if (!game || game.finished) return;
-
-			const attacker = game.players.find((p) => p.index === indexPlayer);
-			const defender = game.players.find((p) => p.index !== indexPlayer);
-
-			if (!attacker || !defender) return;
-			if (game.turn !== attacker.index) return;
-			let shotX = x,
-				shotY = y;
-			if (cmd.type === "randomAttack") {
-				const allCells = [];
-				for (let i = 0; i < 10; i++)
-					for (let j = 0; j < 10; j++) allCells.push([i, j]);
-				game.shots = game.shots || {};
-				const shotCells = Object.keys(game.shots).map((k) =>
-					k.split(",").map(Number)
-				);
-				const available = allCells.filter(
-					([i, j]) => !shotCells.some(([x, y]) => x === i && y === j)
-				);
-				if (available.length) {
-					const [rx, ry] =
-						available[Math.floor(Math.random() * available.length)];
-					shotX = rx;
-					shotY = ry;
-				}
-			}
-			game.shots = game.shots || {};
-			const key = shotX + "," + shotY;
-			if (game.shots[key]) return;
-			game.shots[key] = { by: attacker.index };
-			const result = checkShot(game, defender.index, shotX, shotY);
-			const attackMsg = makeJsonMsg("attack", {
-				position: { x: shotX, y: shotY },
-				currentPlayer: attacker.index,
-				status: result.status,
-			});
-			game.players.forEach((u) => {
-				const wsTarget = players.get(u.name).ws;
-				if (wsTarget && wsTarget.readyState === 1)
-					wsTarget.send(attackMsg);
-			});
-			console.log("Sent:", JSON.parse(attackMsg));
-			if (result.win) {
-				game.finished = true;
-				const finishMsg = makeJsonMsg("finish", { winPlayer: attacker.index });
-				game.players.forEach((u) => {
-					const wsTarget = players.get(u.name).ws;
-					if (wsTarget && wsTarget.readyState === 1)
-						wsTarget.send(finishMsg);
-				});
-				players.get(attacker.name).wins++;
-				sendUpdateWinners();
-				console.log("Sent:", JSON.parse(finishMsg));
-			} else {
-				if (result.status === "miss") {
-					game.turn = defender.index;
-				}
-				sendTurn(game);
-			}
+			handleAttack(cmd, players, games, getShipCells, checkShot, sendTurn, sendUpdateWinners, makeJsonMsg);
 		}
 	});
 	ws.on("close", () => {});
 });
 
 function sendUpdateRoom() {
-	const roomsList = Array.from(rooms.values()).filter(
-		(r) => r.roomUsers && r.roomUsers.length === 1
-	);
+	// Show all rooms, not just those with 1 user
+	const roomsList = Array.from(rooms.values());
 	const data = roomsList.map((room) => ({
 		roomId: room.roomId,
 		roomUsers: room.roomUsers.map((u) => ({
@@ -277,15 +96,21 @@ function checkShot(game, defenderIndex, x, y) {
 		const cells = getShipCells(ship);
 		for (const cell of cells) {
 			if (cell.x === x && cell.y === y) {
-				status = "shot";
 				ship.hits = ship.hits || [];
-				ship.hits.push({ x, y });
+				// Prevent duplicate hits on the same cell
+				const alreadyHit = ship.hits.some(h => h.x === x && h.y === y);
+				if (!alreadyHit) {
+					ship.hits.push({ x, y });
+				}
 				if (ship.hits.length === ship.length) {
 					status = "killed";
+				} else {
+					status = "shot";
 				}
 				break;
 			}
 		}
+		if (status === "shot" || status === "killed") break;
 	}
 	if (ships.every((s) => s.hits && s.hits.length === s.length)) win = true;
 	return { status, win };
@@ -295,8 +120,8 @@ function getShipCells(ship) {
 	const cells = [];
 	for (let i = 0; i < ship.length; i++) {
 		cells.push({
-			x: ship.position.x + (ship.direction ? i : 0),
-			y: ship.position.y + (ship.direction ? 0 : i),
+			x: ship.position.x + (ship.direction ? 0 : i), // direction true = vertical
+			y: ship.position.y + (ship.direction ? i : 0),
 		});
 	}
 	return cells;
